requirements.txt
  # An example requirements file. If your function has other dependencies,
  # add them below
  functions-framework==3.*
  twilio==9.3.7
  google-cloud-logging==3.11.3